+++ src/Ilicop.Web/Controllers/SettingsController.cs
@@ -1,5 +1,4 @@
 ﻿using Geowerkstatt.Ilicop.Web.Contracts;
-using Geowerkstatt.Ilicop.Web.Ilitools;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;