+++ src/Ilicop.Web/IlitoolsEnvironment.cs
@@ -1,4 +1,4 @@
-﻿namespace Geowerkstatt.Ilicop.Web.Ilitools
+﻿namespace Geowerkstatt.Ilicop.Web
 {
     /// <summary>
     /// Runtime-populated metadata about installed ilitools.
@@ -15,6 +15,11 @@
         /// Gets the cache directory.
         /// </summary>
         public string CacheDir { get; init; }
+
+        /// <summary>
+        /// Gets the local model repository directory.
+        /// </summary>
+        public string ModelRepositoryDir { get; init; }
 
         /// <summary>
         /// Indicates whether the GPKG validation should be enabled.
@@ -50,5 +55,25 @@
         /// Indicates whether the ili2gpkg tool is properly setup and initialized.
         /// </summary>
         public bool IsIli2GpkgInitialized => EnableGpkgValidation && !string.IsNullOrWhiteSpace(Ili2GpkgPath);
+
+        /// <inheritdoc/>
+        public override string ToString()
+        {
+            return $$"""
+
+    --------------------------------------------------------------------------
+    ilitools environment:
+    home directory:                   {{HomeDir ?? "unset"}}
+    cache directory:                  {{CacheDir ?? "unset"}}
+    model repository directory:       {{ModelRepositoryDir ?? "unset"}}
+    gpkg validation:                  {{(EnableGpkgValidation ? "enabled" : "disabled")}}
+    ilivalidator version:             {{IlivalidatorVersion ?? "unset"}}
+    ilivalidator initialized:         {{(IsIlivalidatorInitialized ? "yes" : "no")}}
+    ili2gpkg version:                 {{Ili2GpkgVersion ?? "unset"}}
+    ili2gpkg initialized:             {{(IsIli2GpkgInitialized ? "yes" : "no")}}
+    --------------------------------------------------------------------------
+
+    """;
+        }
     }
 }